requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
boto3
passlib[bcrypt]
python-jose 
pydantic[email]
PyJWT
python-multipart
bcrypt<4.1.0